requirements.txt
# Application

Django
psycopg2
psycopg2-binary
static3
gunicorn
dj-database-url
dj-static
python-decouple
django-widget-tweaks
django-extensions
django-auditlog
django-storages
django-extra-fields
boto3
pytz
Pillow

# Django Rest Framework

djangorestframework
itsdangerous
django-cors-headers

# Django Tests

django-test-without-migrations
ipython
